+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Highlights are marked with a pancake 🥞
 - node: Forward events resulting from local actions to app layer [#1290](https://github.com/p2panda/p2panda/pull/1290)
 - node: Cancel processed stream tasks on drop of stream handles [#1300](https://github.com/p2panda/p2panda/pull/1300)
 - spaces: Support partial merge of auth state into a space [#1299](https://github.com/p2panda/p2panda/pull/1299)
+- stream: Move orderer processor from node to stream [#1312](https://github.com/p2panda/p2panda/pull/1312)
 
 ### Changed